Added: Ability to sort the US data for most populous state by population density and area

+++ src/main/java/com/bridgelabz/censusanalyser/IndiaCensusCSV.java
@@ -10,10 +10,10 @@ public class IndiaCensusCSV {
     public int population;
 
     @CsvBindByName(column = "AreaInSqKm")
-    public int totalArea;
+    public long totalArea;
 
     @CsvBindByName(column = "DensityPerSqKm")
-    public int populationDensity;
+    public long populationDensity;
 
     @Override
     public String toString() {
